tp3 done

+++ tp3/include/matrix.hpp
@@ -4,5 +4,5 @@ private:
 public:
     int begin;
     int end;
-    int m[2][2] = {{1, 0}, {0, 1}};
+    long int m[2][2] = {{1, 0}, {0, 1}};
 };
